+++ lua/ulib/server/player.lua
@@ -150,13 +150,12 @@ hook.Add( "CheckPassword", "ULibBanCheck", checkBan, HOOK_LOW )
 		ply - The player to kick.
 		reason - *(Optional)* The reason to give for kicking.
 		calling_ply - *(Optional)* The player doing the kicking.
-		is_ban - *(Optional)* Called from a ban, or just a kick.
 
 	Revisions:
 
 		v2.60 - Fixed a bug with the parameters if you didn't pass reason and calling_ply together.
 ]]
-function ULib.kick( ply, reason, calling_ply, is_ban )
+function ULib.kick( ply, reason, calling_ply )
 	local nick = calling_ply and calling_ply:IsValid() and
 		(string.format( "%s(%s)", calling_ply:Nick(), calling_ply:SteamID() ) or "Console")
 	local steamid = ply:SteamID()
@@ -167,7 +166,7 @@ function ULib.kick( ply, reason, calling_ply, is_ban )
 	else
 		ply:Kick( reason or "[ULX] Kicked from server" )
 	end
-	hook.Call( ULib.HOOK_USER_KICKED, _, steamid, reason or "[ULX] Kicked from server", calling_ply or "Console" )
+	hook.Call( ULib.HOOK_USER_KICKED, _, steamid, reason or "[ULX] Kicked from server", calling_ply )
 end
 
 
@@ -314,7 +313,7 @@ function ULib.unban( steamid, admin )
 	--ULib banlist
 	ULib.bans[ steamid ] = nil
 	ULib.fileWrite( ULib.BANS_FILE, ULib.makeKeyValues( ULib.bans ) )
-	hook.Call( ULib.HOOK_USER_UNBANNED, _, steamid, admin or "Console" )
+	hook.Call( ULib.HOOK_USER_UNBANNED, _, steamid, admin )
 
 end